+++ src/types/vaccineType.ts
@@ -1,3 +1,8 @@
 import { Tables } from "../../database.types";
 
 export type Vaccine = Tables<"vaccine">;
+
+export interface vaccineSchedule extends Omit<Vaccine, 'vaccinae_date' | 'duration'> {
+  startDate: string;
+  endDate: string;
+};
